requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu116
pyyaml==6.0
loguru==0.6.0
pandas==1.5.0
psycopg2-binary==2.9.4
sqlalchemy==1.4.41
tqdm==4.64.1
torch
torchvision
torchaudio
transformers
scikit-learn==1.1.2
nltk==3.7
catboost==1.1
optuna==3.0.3
joblib==1.2.0